Pass.jl

Pass is a Julia interface to the unix pass command. This package provides PassStore, a read-only Dict-like object inspired by Base.EnvDict.

Installation

Pass.jl can be installed using Pkg.

pkg> add Pass

Installation of Pass.jl does not attempt to install pass (or gpg) on your system. pass needs to be available at the command line in order for Pass.jl to operate successfully.

Example Usage

As an example, assume that you've previously stored an API key for a service in the default location for pass.

$ echo 'EXAMPLE-API-KEY' | pass insert -e service/api_key

Warning

Running the command above puts EXAMPLE-API-KEY into your history, which stores your secrets on disk in plaintext. For production use, it would be better to store the key interactively.

pass insert service/api_key
<...>

If configured correctly, this key can be retrieved at the command line.

$ pass show service/api_key
EXAMPLE-API-KEY

From Julia, Pass.jl provides access to the same information.

using Pass
store = PassStore()  # default `pass` location
api_key = store["service/api_key"]  # "EXAMPLE-API-KEY"

Tip

For ease of use, you may want to add your default PassStore to startup.jl.

using Pass
const PASS = PassStore()

PASS can be used just like ENV to retrieve secrets.

api_key = PASS["service/api_key"]

Pass.jl does not cache or store any of the values, for both simplicity and security.

API

Pass.Pass

Pass.PassStore

Pass.PassStore — Type

PassStore(dir=nothing)

A password store interface that provides dictionary-like access to the pass command-line password manager.

Arguments

dir::Union{String,Nothing}: The password store directory path. If nothing, uses the default store location (~/.password-store).

Examples

# Use default password store
store = PassStore()

# Use custom password store directory
store = PassStore("/path/to/my/store")

# Retrieve passwords
password = store["service/password"]

Throws

SystemError: If the pass command is not available

ArgumentError: If the specified directory doesn't exist or isn't initialized

Pass.Pass — Module

Provides PassStore, a Dict-like interface to the unix pass command

Setup

Danger

Pass.jl is only as secure as your pass configuration, which is your responsibility to manage. The examples here and in the unit tests are meant to serve as a starting point for moving away from plaintext credential storage. Review man gpg and man pass to learn about the options that are right for you. Note that the example configuration in the unit tests are only meant to

Approximate steps:

Generate a new gpg key gpg --full-generate-key

Show the id for the generated key gpg --list-keys --keyid-format LONG

Initialize the password store pass init <KEYID>

Insert a secret pass insert <path-to-secret>

Retrieve the secret pass show <path-to-secret>